loaded js files as module in html, import and export general.js and project.js and index.js. Also moved a bit forward with the TOher projects section dynamic DOM manipulation, this is stil in progress and giving errors
