+++ Contact/contact.html
@@ -59,6 +59,6 @@
         </ul>
     </footer>
     <script src="/ironhack-midterm-project/Contact/contact.js"></script>
-    <script src="/ironhack-midterm-project/Utils/general.js"></script>
+    <script type="module" src="/ironhack-midterm-project/Utils/general.js"></script>
 </body>
 </html>
