+++ Project/project.html
@@ -38,21 +38,21 @@
         <h3>Other Projects</h3>
         <div>
             <article>
-                <img src="/ironhack-midterm-project/Assets/Projects/1.jpg" alt="Background image with colors">
-                <p id="titleArticle" class="body-medium">Orbit</p>
-                <p id="descriptionArticle" class="headline-regular">UI Design</p>
+                <img alt="Background image with colors">
+                <p id="titleArticle" class="body-medium"></p>
+                <p id="descriptionArticle" class="headline-regular"></p>
+                <button class="secondaryButton headline-regular">Learn More</button>
+            </article>
+            <article>
+                <img alt="Background image with colors">
+                <p class="body-medium"></p>
+                <p class="headline-regular"></p>
                 <button class="secondaryButton headline-regular">Learn more</button>
             </article>
             <article>
-                <img src="/ironhack-midterm-project/Assets/Projects/2.jpg" alt="Background image with colors">
-                <p class="body-medium">Purify</p>
-                <p class="headline-regular">UI/UX Design</p>
-                <button class="secondaryButton headline-regular">Learn more</button>
-            </article>
-            <article>
-                <img src="/ironhack-midterm-project/Assets/Projects/3.jpg" alt="Background image with colors">
-                <p class="body-medium">CryptoPie</p>
-                <p class="headline-regular">App Development</p>
+                <img alt="Background image with colors">
+                <p class="body-medium"></p>
+                <p class="headline-regular"></p>
                 <button class="secondaryButton headline-regular">Learn more</button>
             </article>
         </div>
@@ -83,7 +83,7 @@
             <li>Privacy Policy</li>
         </ul>
     </footer>
-    <script src="/ironhack-midterm-project/Project/project.js"></script>
-    <script src="/ironhack-midterm-project/Utils/general.js"></script>
+    <script type="module" src="/ironhack-midterm-project/Project/project.js"></script>
+    <script type="module" src="/ironhack-midterm-project/Utils/general.js"></script>
 </body>
 </html>
