+++ index.html
@@ -153,7 +153,7 @@
             <li>Privacy Policy</li>
         </ul>
     </footer>
-    <script src="/ironhack-midterm-project/index.js"></script>
-    <script src="/ironhack-midterm-project/Utils/general.js"></script>
+    <script type="module" src="/ironhack-midterm-project/index.js"></script>
+    <script type="module" src="/ironhack-midterm-project/Utils/general.js"></script>
 </body>
 </html>
